fix some tests
work in progress
this is something I need to check up on
with someone who is more up to date with Angular
should a service spec test provide the service?
what does this even mean?

the test passes now though
still is this the way to do it?
will check later
#TODO T_T